+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0] - 2024-02-13
+
+### Added
+
+* New option for the `TransactionHelper` where the rollback properties can be set.
+* Default logic for whitelist and blacklist is created for rollback.
+* Interception is not used if there is a transaction active and new transaction is not requested.
+
 ## [1.12.3] - 2024-01-11
 
 ### Fixed